メモの入力をするwriteMemoFromInterfaceを実装しました、
